Split itest-karaf so that only Weld tests are run with Weld profiles and vice versa
