+++ itest/src/it/itest-karaf/src/test/java/org/ops4j/pax/cdi/test/karaf/weld/PaxCdiDeltaSpikeTest.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.ops4j.pax.cdi.test.karaf;
+package org.ops4j.pax.cdi.test.karaf.weld;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.notNullValue;
